pyproject.toml
[project]
name = "arc-lang-public"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "asyncpg>=0.31.0",
    "devtools>=0.12.2",
    "google-genai>=1.54.0",
    "logfire>=4.16.0",
    "matplotlib>=3.10.1",
    "openai>=2.9.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.28.0",
    "python-dotenv>=1.1.0",
    "xai-sdk>=1.5.0",
]

[dependency-groups]
dev = ["ruff>=0.11.4"]


[tool.mypy]
plugins = 'pydantic.mypy'
strict = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_untyped_decorators = false
disallow_subclassing_any = false
ignore_missing_imports = true
follow_imports = 'skip'
exclude = ["venv", ".venv", "alembic", "app/dbs/sql_gen/gen/db.py"]

[tool.ruff]
target-version = "py312"
exclude = [
    "alembic",
    "app/dbs/sql_gen/gen/db.py",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
unfixable = [
    "F401", # unused imports
]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
